requirements.txt
dnspython
xmpppy
requests
flask
flask-login
python-ldap
blessings
